+++ animation_nodes/nodes/spline/spline_from_gp_stroke.py
@@ -2,17 +2,16 @@ import bpy
 from ... data_structures import PolySpline, FloatList
 from ... base_types import AnimationNode, VectorizedSocket
 
-class SplinesFromGPStrokesNode(bpy.types.Node, AnimationNode):
-    bl_idname = "an_SplinesFromGPStrokesNode"
-    bl_label = "Splines From GP Strokes"
+class SplineFromGPStrokeNode(bpy.types.Node, AnimationNode):
+    bl_idname = "an_SplineFromGPStrokeNode"
+    bl_label = "Spline From GP Stroke"
     codeEffects = [VectorizedSocket.CodeEffect]
 
     useStrokeList: VectorizedSocket.newProperty()
 
     def create(self):
-        socket = self.newInput(VectorizedSocket("GPStroke", "useStrokeList",
+        self.newInput(VectorizedSocket("GPStroke", "useStrokeList",
             ("Stroke", "stroke"), ("Strokes", "strokes")))
-        socket.dataIsModified = True
 
         self.newOutput(VectorizedSocket("Spline", "useStrokeList",
             ("Spline", "spline"), ("Splines", "splines")))
@@ -22,5 +21,4 @@ class SplinesFromGPStrokesNode(bpy.types.Node, AnimationNode):
 
     def splinesFromStrokes(self, stroke):
         if stroke is None: return PolySpline()
-        radii = FloatList.fromValues(stroke.pressures)
-        return PolySpline(points = stroke.vertices, radii = radii, cyclic = stroke.drawCyclic)
+        return PolySpline(points = stroke.vertices, radii = stroke.pressures, cyclic = stroke.drawCyclic)